info_core/requirements.txt
asgiref==3.8.1
Django==5.0.8
django-cuser==2017.3.16
django-modeltranslation==0.19.7
django-tinymce==4.1.0
pillow==10.4.0
python-gettext==5.0
sqlparse==0.5.1
tzdata==2024.1
